Fixed size cast warning

+++ src/rhash/hash_disc.c
@@ -1219,7 +1219,7 @@ static int rc_hash_wii_disc(md5_state_t* md5, const rc_hash_iterator_t* iterator
       }
     }
     else { /* Decrypted */
-      if (rc_hash_nintendo_disc_partition(md5, iterator, file_handle, part_offset, 2) == 0) {
+      if (rc_hash_nintendo_disc_partition(md5, iterator, file_handle, (uint32_t)part_offset, 2) == 0) {
         return rc_hash_iterator_error(iterator, "Failed to hash Wii partition");
       }
     }
